// ma/node/cdk/stacks/batch/models/BatchJobDef.ts
import BatchRepo from '../enums/BatchRepo';

interface BatchJobDef {
  name: string;
  image: BatchRepo;
  imageTag?: string;
  vcpu: number;
  memory: number;
  retryAttempts: number;
  timeoutSeconds: number;
  command: string[];
  environment?: {
    [key: string]: string;
  };
  parameters?: {
    [key: string]: string;
  };
}

export default BatchJobDef;